Refactoring SearchTextForLocalization 2
