+++ Assets/AutoTranslate/Editor/Search Text  Scene/StatusLocalizationScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76d3351deee01ee48aac227efdb01aaa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
