+++ Assets/AutoTranslate/Editor/Translate Window/LocalizeStringEventExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8f198f600369be48bc07a53fe0cddf9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
